Extract CreateExpenseDialogView to its own file and split spec into tests directory
- Move CreateExpenseDialogView and CreateExpenseDialogViewProps out of
  CreateExpenseDialog.tsx into CreateExpenseDialogView.tsx (view has its own
  stories and tests, which requires a separate file per CLAUDE.md)
- Split CreateExpenseDialog.spec.tsx (389 lines) into
  CreateExpenseDialog-tests/ with two domain-specific files:
  CreateExpenseDialogView.spec.tsx (view tests) and
  CreateExpenseDialog.spec.tsx (container integration tests)
- Update stories and container imports to reference the new file locations
