+++ src/components/reconcile/CreateExpenseDialogView.tsx
@@ -1,0 +1,214 @@
+"use client";
+
+import { useId } from "react";
+
+import { Button } from "@/components/ui/button";
+import {
+  Dialog,
+  DialogContent,
+  DialogFooter,
+  DialogHeader,
+  DialogTitle,
+} from "@/components/ui/dialog";
+import { ReconciliationExpenseType } from "@/lib/firebase/schema/reconciliation-expenses";
+
+import { CREATE_EXPENSE_DIALOG_COPY } from "./CreateExpenseDialog.copy";
+
+export interface CreateExpenseDialogViewProps {
+  open: boolean;
+  onOpenChange: (open: boolean) => void;
+  name: string;
+  onNameChange: (value: string) => void;
+  type: ReconciliationExpenseType;
+  onTypeChange: (value: ReconciliationExpenseType) => void;
+  typicalAmount: string;
+  onTypicalAmountChange: (value: string) => void;
+  nameError: string | undefined;
+  typicalAmountError: string | undefined;
+  submitError: string | undefined;
+  onSubmit: () => void;
+  isSubmitting: boolean;
+}
+
+export function CreateExpenseDialogView({
+  open,
+  onOpenChange,
+  name,
+  onNameChange,
+  type,
+  onTypeChange,
+  typicalAmount,
+  onTypicalAmountChange,
+  nameError,
+  typicalAmountError,
+  submitError,
+  onSubmit,
+  isSubmitting,
+}: CreateExpenseDialogViewProps) {
+  const baseId = useId();
+  const nameInputId = `${baseId}-name`;
+  const nameErrorId = `${baseId}-name-error`;
+  const amountInputId = `${baseId}-amount`;
+  const amountErrorId = `${baseId}-amount-error`;
+  const typeLabelId = `${baseId}-type-label`;
+
+  return (
+    <Dialog
+      open={open}
+      onOpenChange={(isOpen: boolean) => {
+        if (!isOpen && isSubmitting) return;
+        onOpenChange(isOpen);
+      }}
+    >
+      <DialogContent>
+        <DialogHeader>
+          <DialogTitle>{CREATE_EXPENSE_DIALOG_COPY.title}</DialogTitle>
+        </DialogHeader>
+        <form
+          noValidate
+          onSubmit={(e) => {
+            e.preventDefault();
+            onSubmit();
+          }}
+        >
+          <div className="flex flex-col gap-4 py-2">
+            {/* Name field */}
+            <div className="flex flex-col gap-1.5">
+              <label
+                htmlFor={nameInputId}
+                className="text-sm font-medium leading-none"
+              >
+                {CREATE_EXPENSE_DIALOG_COPY.nameLabel}
+              </label>
+              <input
+                id={nameInputId}
+                type="text"
+                value={name}
+                onChange={(e) => {
+                  onNameChange(e.target.value);
+                }}
+                placeholder={CREATE_EXPENSE_DIALOG_COPY.namePlaceholder}
+                aria-invalid={nameError !== undefined}
+                aria-describedby={nameError ? nameErrorId : undefined}
+                className="flex h-8 w-full rounded-lg border border-input bg-background px-3 py-1 text-sm ring-offset-background placeholder:text-muted-foreground focus-visible:outline-none focus-visible:ring-2 focus-visible:ring-ring disabled:cursor-not-allowed disabled:opacity-50"
+              />
+              {nameError !== undefined && (
+                <p
+                  id={nameErrorId}
+                  role="alert"
+                  className="text-xs text-destructive"
+                >
+                  {nameError}
+                </p>
+              )}
+            </div>
+
+            {/* Type toggle */}
+            <div className="flex flex-col gap-1.5">
+              <p id={typeLabelId} className="text-sm font-medium leading-none">
+                {CREATE_EXPENSE_DIALOG_COPY.typeLabel}
+              </p>
+              <div
+                role="group"
+                aria-labelledby={typeLabelId}
+                className="flex gap-2"
+              >
+                <button
+                  type="button"
+                  aria-pressed={
+                    type === ReconciliationExpenseType.StatementBalance
+                  }
+                  onClick={() => {
+                    onTypeChange(ReconciliationExpenseType.StatementBalance);
+                  }}
+                  className={`rounded-lg border px-3 py-1.5 text-sm font-medium transition-colors ${
+                    type === ReconciliationExpenseType.StatementBalance
+                      ? "border-primary bg-primary text-primary-foreground"
+                      : "border-border bg-background hover:bg-muted"
+                  }`}
+                >
+                  {CREATE_EXPENSE_DIALOG_COPY.typeStatementBalance}
+                </button>
+                <button
+                  type="button"
+                  aria-pressed={
+                    type === ReconciliationExpenseType.RunningBalance
+                  }
+                  onClick={() => {
+                    onTypeChange(ReconciliationExpenseType.RunningBalance);
+                  }}
+                  className={`rounded-lg border px-3 py-1.5 text-sm font-medium transition-colors ${
+                    type === ReconciliationExpenseType.RunningBalance
+                      ? "border-primary bg-primary text-primary-foreground"
+                      : "border-border bg-background hover:bg-muted"
+                  }`}
+                >
+                  {CREATE_EXPENSE_DIALOG_COPY.typeRunningBalance}
+                </button>
+              </div>
+            </div>
+
+            {/* Monthly amount field */}
+            <div className="flex flex-col gap-1.5">
+              <label
+                htmlFor={amountInputId}
+                className="text-sm font-medium leading-none"
+              >
+                {CREATE_EXPENSE_DIALOG_COPY.amountLabel}
+              </label>
+              <input
+                id={amountInputId}
+                type="number"
+                min="0.01"
+                step="0.01"
+                value={typicalAmount}
+                onChange={(e) => {
+                  onTypicalAmountChange(e.target.value);
+                }}
+                placeholder={CREATE_EXPENSE_DIALOG_COPY.amountPlaceholder}
+                aria-invalid={typicalAmountError !== undefined}
+                aria-describedby={
+                  typicalAmountError ? amountErrorId : undefined
+                }
+                className="flex h-8 w-full rounded-lg border border-input bg-background px-3 py-1 text-sm ring-offset-background placeholder:text-muted-foreground focus-visible:outline-none focus-visible:ring-2 focus-visible:ring-ring disabled:cursor-not-allowed disabled:opacity-50"
+              />
+              {typicalAmountError !== undefined && (
+                <p
+                  id={amountErrorId}
+                  role="alert"
+                  className="text-xs text-destructive"
+                >
+                  {typicalAmountError}
+                </p>
+              )}
+            </div>
+
+            {submitError !== undefined && (
+              <p role="alert" className="text-xs text-destructive">
+                {submitError}
+              </p>
+            )}
+          </div>
+
+          <DialogFooter className="mt-2">
+            <Button
+              type="button"
+              variant="outline"
+              onClick={() => {
+                onOpenChange(false);
+              }}
+              disabled={isSubmitting}
+            >
+              {CREATE_EXPENSE_DIALOG_COPY.cancelButton}
+            </Button>
+            <Button type="submit" disabled={isSubmitting}>
+              {isSubmitting
+                ? CREATE_EXPENSE_DIALOG_COPY.creatingButton
+                : CREATE_EXPENSE_DIALOG_COPY.submitButton}
+            </Button>
+          </DialogFooter>
+        </form>
+      </DialogContent>
+    </Dialog>
+  );
+}
